Wait much longer when receiving a OVER_QUERY_LIMIT.
Our logs are full of this message.  The 0.01 second delay is not doing
much (the original author thought it was 0.1 seconds, and lists the time
configured as "ms" when it's actually microseconds).

Bringing in line with the google suggestions here:
https://developers.google.com/maps/documentation/business/articles/usage_limits

Will wait only 2 seconds, then try again.  After a few failures, an
error message stating that we are over the daily query limit is written
and no more queries are sent for this instantiation.